+++ clu/data/dataset_iterator.py
@@ -31,10 +31,8 @@ import abc
 import collections.abc
 import concurrent.futures
 import dataclasses
-import os
 import threading
-import typing
-from typing import Any, Mapping, Optional, Sequence, Tuple, TypeVar, Union
+from typing import Mapping, Optional, Sequence, Tuple, TypeVar, Union
 
 from absl import logging
 from clu import asynclib
@@ -131,81 +129,7 @@ class DatasetIterator(collections.abc.Iterator):  # pytype: disable=ignored-abst
     return self.restore(filename)
 
 
-class TfDatasetIterator(DatasetIterator):
-  """DatasetIterator for wrapping a `tf.data.Dataset`."""
 
-  def __init__(self, dataset, *, checkpoint: bool):
-    """Wraps `tf.data.Dataset` object into the `DatasetIterator` interface.
-
-    Warning: Do not wrap this interator to do asynchronous prefetching if you
-    use `checkpoint=True` (default). tf.data iterators must be saved()
-    synchronously.
-
-    Args:
-      dataset: The dataset to wrap. Elements are converted to NumPy arrays but
-        no additional prefetching is done. tf.data should automatically prefetch
-        elements (to CPU memory).
-      checkpoint: Whether to checkpoint the dataset iterator object.
-        Checkpointing dataset iterators is required for handling job
-        pre-emptions but depending on your input pipeline can result in very
-        large checkpoints. If set to False save() and load() are no-ops.
-    """
-    try:
-      # Since this is the only class in this module using TF we only import
-      # tensorflow if needed.
-      if typing.TYPE_CHECKING:
-        tf = Any
-      else:
-        import tensorflow as tf  # pylint: disable=g-import-not-at-top
-    except ImportError as e:
-      raise RuntimeError("When using TfDatasetIterator your binary must "
-                         "depend on //third_party/py/tensorflow.") from e
-    self._tf = tf
-
-    if not isinstance(dataset, tf.data.Dataset):
-      raise ValueError("`dataset` must be an instance of `tf.data.Dataset` "
-                       f"but got {type(dataset)}.")
-    self._dataset = dataset
-    self._checkpoint = checkpoint
-    assert self.element_spec  # Verify element spec.
-    self.iterator = iter(dataset)
-    self._ckpt = tf.train.Checkpoint(ds=self.iterator)
-
-  def get_next(self) -> Element:
-    return next(self)
-
-  def __next__(self) -> Element:
-    return {k: np.asarray(v) for k, v in next(self.iterator).items()}
-
-  def reset(self):
-    self.iterator = iter(self._dataset)
-    self._ckpt = self._tf.train.Checkpoint(ds=self.iterator)
-
-  @property
-  def element_spec(self) -> ElementSpec:
-    element_spec = self._dataset.element_spec
-    if not isinstance(element_spec, dict):
-      raise ValueError("Dataset elements must be flat dictionaries but got "
-                       f"{element_spec}.")
-    invalid_features = [
-        k for k, v in element_spec.items()
-        if not isinstance(v, self._tf.TensorSpec)
-    ]
-    if invalid_features:
-      raise ValueError(f"Features {invalid_features} are not tensors. Dataset "
-                       "elements must be flat dictionaries of tensors.")
-    return {
-        k: ArraySpec(dtype=v.dtype.as_numpy_dtype, shape=tuple(v.shape))
-        for k, v in element_spec.items()
-    }
-
-  def save(self, filename: epath.Path):
-    if self._checkpoint:
-      self._ckpt.write(os.fspath(filename))
-
-  def restore(self, filename: epath.Path):
-    if self._checkpoint:
-      self._ckpt.read(os.fspath(filename)).assert_consumed()
 
 
 class PeekableDatasetIterator(DatasetIterator):